refactoring early stop test code
